# contracts/agent-api/agent-openapi-v1.yaml
openapi: 3.1.0
info:
  title: AIResearcher Agent API
  version: 1.0.0
  description: >-
    Python Agent 向 Java BFF 暴露的本地论文原件库与 RAG API。Java 是唯一调用方；
    原件、论文、任务、检索和模型运行数据均由 Python Agent 管理。
...
servers:
  - url: /
    description: Agent 服务根地址
...
paths:
  /agent-api/v1/library:
    get:
      tags: [Library]
      operationId: getAgentLibrary
      summary: 获取本地论文原件库状态
      parameters:
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/LibraryResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/library/files:
    get:
      tags: [Library]
      operationId: listAgentLibraryFiles
      summary: 分页列出论文原件及其知识库对应状态
      description: 原件清单独立于知识库论文；NOT_INGESTED 原件尚未创建 Paper 或入库任务。
      parameters:
        - $ref: '#/components/parameters/LibraryFilesOffset'
        - $ref: '#/components/parameters/LibraryFilesLimit'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/LibraryFilesResponse'
        '400':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
    post:
      tags: [Library]
      operationId: uploadAgentLibraryFile
      summary: 上传一篇 PDF 原件但不创建入库任务
      description: >-
        multipart 请求必须且只能包含一个名为 file 的 PDF，最大 50 MB。文件完成安全校验、
        SHA-256 去重和原子落盘后仅登记为原件；客户端必须另行调用 ingestion 接口才会解析和建库。
      parameters:
        - $ref: '#/components/parameters/RequiredRequestId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 单个 application/pdf 原件。
      responses:
        '200':
          $ref: '#/components/responses/LibraryFileUploadResponse'
        '400':
          $ref: '#/components/responses/AgentErrorResponse'
        '413':
          $ref: '#/components/responses/AgentErrorResponse'
        '415':
          $ref: '#/components/responses/AgentErrorResponse'
        '422':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/library/files/{libraryFileId}/file:
    get:
      tags: [Library]
      operationId: getAgentLibraryFile
      summary: 预览或按 Range 读取已登记 PDF 原件
      parameters:
        - $ref: '#/components/parameters/LibraryFileId'
        - $ref: '#/components/parameters/RequiredRequestId'
        - $ref: '#/components/parameters/OptionalRange'
      responses:
        '200':
          $ref: '#/components/responses/PdfFile'
        '206':
          $ref: '#/components/responses/PdfPartialFile'
        '400':
          $ref: '#/components/responses/AgentErrorResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '409':
          $ref: '#/components/responses/AgentErrorResponse'
        '416':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/library/files/{libraryFileId}/ingestion:
    post:
      tags: [Library]
      operationId: ingestAgentLibraryFile
      summary: 手动将一篇可用原件录入知识库
      description: >-
        创建 Paper 和后台入库任务；相同 SHA-256 已有 Paper 时复用既有论文和任务。
        MISSING 或 REPLACED 原件返回冲突，扫描和上传本身绝不调用此操作。
      parameters:
        - $ref: '#/components/parameters/LibraryFileId'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/LibraryFileIngestionResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '409':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/library/scans:
    post:
      tags: [Library]
      operationId: createAgentLibraryScan
      summary: 创建一次后台原件库扫描
      description: >-
        同时只允许一个 QUEUED 或 RUNNING 的扫描任务。扫描只登记和核对原件，
        不创建 Paper、入库任务、chunk 或向量。
      parameters:
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '202':
          description: 扫描任务已排队。
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryScan'
              example:
                scanId: scan-demo-001
                status: QUEUED
                discoveredCount: 0
                registeredCount: 0
                unchangedCount: 0
                duplicateCount: 0
                excludedCount: 0
                skippedCount: 0
                failedCount: 0
                createdAt: '2026-08-29T00:00:00Z'
                startedAt: null
                completedAt: null
                failure: null
        '409':
          description: 已有活动扫描；错误 code 固定为 LIBRARY_SCAN_ACTIVE。
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentApiError'
              example:
                schemaVersion: '1.0'
                code: LIBRARY_SCAN_ACTIVE
                message: 原件库扫描正在进行。
                requestId: req-library-scan-001
                retryable: true
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/library/scans/{scanId}:
    get:
      tags: [Library]
      operationId: getAgentLibraryScan
      summary: 获取原件库扫描任务
      parameters:
        - $ref: '#/components/parameters/ScanId'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/LibraryScanResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/library/scans/{scanId}/items:
    get:
      tags: [Library]
      operationId: listAgentLibraryScanItems
      summary: 分页获取扫描文件结果
      parameters:
        - $ref: '#/components/parameters/ScanId'
        - $ref: '#/components/parameters/ScanItemsOffset'
        - $ref: '#/components/parameters/ScanItemsLimit'
        - $ref: '#/components/parameters/ScanItemOutcomeFilter'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/LibraryScanItemsResponse'
        '400':
          $ref: '#/components/responses/AgentErrorResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/papers:
    post:
      tags: [Papers]
      operationId: uploadAgentPaper
      summary: 上传一篇文本型 PDF 并创建入库任务
      deprecated: true
      description: >-
        兼容性入口；新客户端应依次调用 library/files 上传原件和 ingestion 手动入库。
        multipart 请求必须且只能包含一个名为 file 的 PDF，最大 50 MB、500 页。
        相同 SHA-256 返回原论文和任务并将 duplicate 置为 true，不创建第二份文件或向量。
      parameters:
        - $ref: '#/components/parameters/RequiredRequestId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 单个 application/pdf 文件。
      responses:
        '200':
          description: 已接收新论文，或返回 SHA-256 相同的既有论文。
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperUploadResponse'
              example:
                paper:
                  paperId: paper-demo-001
                  title: Synthetic Bilingual Research
                  authors: [Demo Author]
                  publicationYear: 2026
                  fileName: synthetic-demo.pdf
                  fileSizeBytes: 4096
                  libraryRelativePath: synthetic-demo.pdf
                  sourceStatus: AVAILABLE
                  status: PROCESSING
                  searchable: false
                  pageCount: null
                  createdAt: '2026-01-01T00:00:00Z'
                  updatedAt: '2026-01-01T00:00:00Z'
                  currentIngestion:
                    jobId: job-demo-001
                    status: QUEUED
                    stage: QUEUED
                    attempt: 0
                    maxAttempts: 3
                    canRetry: false
                    failure: null
                ingestionJob:
                  jobId: job-demo-001
                  paperId: paper-demo-001
                  status: QUEUED
                  stage: QUEUED
                  attempt: 0
                  maxAttempts: 3
                  canRetry: false
                  failure: null
                  createdAt: '2026-01-01T00:00:00Z'
                  startedAt: null
                  completedAt: null
                duplicate: false
        '400':
          $ref: '#/components/responses/AgentErrorResponse'
        '413':
          $ref: '#/components/responses/AgentErrorResponse'
        '415':
          $ref: '#/components/responses/AgentErrorResponse'
        '422':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
    get:
      tags: [Papers]
      operationId: listAgentPapers
      summary: 列出默认知识库中的论文
      parameters:
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          description: 论文列表，按创建时间倒序。
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperListResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/papers/{paperId}:
    get:
      tags: [Papers]
      operationId: getAgentPaper
      summary: 获取论文详情
      parameters:
        - $ref: '#/components/parameters/PaperId'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/PaperResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/papers/{paperId}/exclusion:
    post:
      tags: [Papers]
      operationId: excludeAgentPaper
      summary: 将论文移出知识库但保留原件
      parameters:
        - $ref: '#/components/parameters/PaperId'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/PaperResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '409':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
    delete:
      tags: [Papers]
      operationId: restoreAgentPaper
      summary: 重新加入原件可用的已排除论文
      parameters:
        - $ref: '#/components/parameters/PaperId'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/PaperResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '409':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/papers/{paperId}/file:
    get:
      tags: [Papers]
      operationId: getAgentPaperFile
      summary: 下载或按 Range 读取原始 PDF
      parameters:
        - $ref: '#/components/parameters/PaperId'
        - $ref: '#/components/parameters/RequiredRequestId'
        - $ref: '#/components/parameters/OptionalRange'
      responses:
        '200':
          $ref: '#/components/responses/PdfFile'
        '206':
          $ref: '#/components/responses/PdfPartialFile'
        '400':
          $ref: '#/components/responses/AgentErrorResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '416':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/ingestion-jobs/{jobId}:
    get:
      tags: [Ingestion jobs]
      operationId: getAgentIngestionJob
      summary: 获取入库任务状态和阶段
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/IngestionJobResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/ingestion-jobs/{jobId}/retry:
    post:
      tags: [Ingestion jobs]
      operationId: retryAgentIngestionJob
      summary: 重试一个可重试的失败任务
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/RequiredRequestId'
      responses:
        '200':
          $ref: '#/components/responses/IngestionJobResponse'
        '404':
          $ref: '#/components/responses/AgentErrorResponse'
        '409':
          $ref: '#/components/responses/AgentErrorResponse'
        '500':
          $ref: '#/components/responses/AgentErrorResponse'
        '503':
          $ref: '#/components/responses/AgentErrorResponse'
  /agent-api/v1/conversations/{conversationId}/messages/stream:
    post:
      tags: [Conversations]
      operationId: streamAgentConversationMessage
      summary: 使用原生 Tool Calling 流式生成助手消息
      description: >-
        Java 必须传入有效 X-Request-Id。HTTP 200 表示流已建立；建流前非 2xx
        使用 StreamOpenError。建流后的失败只能由唯一 run.failed 终止事件表达。
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/RequiredRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatStreamRequest'
            examples:
              defaultKnowledgeBase:
                value:
                  content: 论文中的核心实验结论是什么？
                  paperIds: []
              selectedPaper:
                value:
                  content: 第二页报告了什么结果？
                  paperIds: [paper-demo-001]
      responses:
        '200':
          $ref: '#/components/responses/SseStream'
        '400':
          $ref: '#/components/responses/BadStreamRequest'
        '500':
          $ref: '#/components/responses/InternalStreamError'
        '503':
          $ref: '#/components/responses/StreamServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedStreamOpenError'
      x-sse-event-schema:
        $ref: '#/components/schemas/SseEvent'
      x-sse-contract: ../sse/v1/README.md
components:
  parameters:
...
    LibraryFileId:
      name: libraryFileId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
    ScanId:
      name: scanId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
    LibraryFilesOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    LibraryFilesLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 100
    RequiredRequestId:
      name: X-Request-Id
      in: header
      required: true
      description: Java 接收或生成并原样转发的请求追踪标识。
      schema:
        $ref: '#/components/schemas/Identifier'
    OptionalRange:
      name: Range
      in: header
      required: false
      description: 单一 HTTP bytes range；不支持 multipart ranges。
      schema:
        type: string
        pattern: '^bytes=(?:[0-9]+-[0-9]*|-[0-9]+)$'
        example: bytes=0-1023
  headers:
    XRequestId:
      description: 本次请求的有效请求追踪标识。
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
  schemas:
...
    LibraryScan:
      type: object
      additionalProperties: false
      required:
        - scanId
        - status
        - discoveredCount
        - registeredCount
        - unchangedCount
        - duplicateCount
        - excludedCount
        - skippedCount
        - failedCount
        - createdAt
        - startedAt
        - completedAt
        - failure
      properties:
        scanId:
          $ref: '#/components/schemas/Identifier'
        status:
          $ref: '#/components/schemas/LibraryScanStatus'
        discoveredCount: { type: integer, minimum: 0 }
        registeredCount:
          type: integer
          minimum: 0
          description: 本次新增登记的原件数量；登记原件不会自动创建 Paper、入库任务或向量。
        unchangedCount: { type: integer, minimum: 0 }
        duplicateCount: { type: integer, minimum: 0 }
        excludedCount: { type: integer, minimum: 0 }
        skippedCount: { type: integer, minimum: 0 }
        failedCount: { type: integer, minimum: 0 }
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        failure:
          oneOf:
            - $ref: '#/components/schemas/LibraryScanFailure'
            - type: 'null'
...
    AgentApiError:
      type: object
      additionalProperties: false
      required: [schemaVersion, code, message, requestId, retryable]
      properties:
        schemaVersion:
          const: '1.0'
        code:
          type: string
          minLength: 1
          maxLength: 128
        message:
          type: string
          minLength: 1
          maxLength: 2048
        requestId:
          $ref: '#/components/schemas/Identifier'
        retryable:
          type: boolean
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, reason]
            properties:
              field:
                type: string
                minLength: 1
                maxLength: 256
              reason:
                type: string
                minLength: 1
                maxLength: 1024
...
    SseEvent:
      $ref: ../sse/v1/sse-event.schema.json
...
  responses:
    LibraryResponse:
      description: 本地论文原件库状态。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LibraryInfo'
          example:
            rootPath: 'D:/AIResearcher/.private/paper-library'
            supportedExtensions: ['.pdf']
            scanInProgress: false
            latestScan: null
    LibraryFilesResponse:
      description: 分页原件与知识库对应状态。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LibraryFilesPage'
          example:
            items:
              - libraryFileId: library-file-demo-001
                relativePath: uploads/synthetic-demo.pdf
                fileName: synthetic-demo.pdf
                fileSizeBytes: 4096
                sha256: 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
                sourceStatus: AVAILABLE
                knowledgeStatus: NOT_INGESTED
                paperId: null
                paperTitle: null
                searchable: false
                currentIngestion: null
                discoveredAt: '2026-08-29T00:00:00Z'
                lastSeenAt: '2026-08-29T00:00:00Z'
                updatedAt: '2026-08-29T00:00:00Z'
            total: 1
            offset: 0
            limit: 100
    LibraryFileUploadResponse:
      description: PDF 原件已登记，但未创建 Paper 或入库任务。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LibraryFileUploadResponse'
          example:
            libraryFile:
              libraryFileId: library-file-demo-001
              relativePath: uploads/synthetic-demo.pdf
              fileName: synthetic-demo.pdf
              fileSizeBytes: 4096
              sha256: 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
              sourceStatus: AVAILABLE
              knowledgeStatus: NOT_INGESTED
              paperId: null
              paperTitle: null
              searchable: false
              currentIngestion: null
              discoveredAt: '2026-08-29T00:00:00Z'
              lastSeenAt: '2026-08-29T00:00:00Z'
              updatedAt: '2026-08-29T00:00:00Z'
            duplicate: false
    LibraryFileIngestionResponse:
      description: 原件已关联 Paper，并返回当前入库任务。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LibraryFileIngestionResponse'
    LibraryScanResponse:
      description: 原件库扫描任务详情。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LibraryScan'
...
    AgentErrorResponse:
      description: Agent REST 错误；Java 必须映射为稳定 Web Result 错误。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AgentApiError'
          example:
            schemaVersion: '1.0'
            code: PAPER_NOT_FOUND
            message: 未找到指定论文。
            requestId: req-demo-error-001
            retryable: false
    PdfFile:
      description: 完整 PDF；不使用 JSON 包装。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Accept-Ranges:
          $ref: '#/components/headers/AcceptRanges'
        Content-Length:
          $ref: '#/components/headers/ContentLength'
        ETag:
          $ref: '#/components/headers/ETag'
      content:
        application/pdf:
          schema:
            type: string
            format: binary
    PdfPartialFile:
      description: 单一 Range 的 PDF 字节片段；不使用 JSON 包装。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Accept-Ranges:
          $ref: '#/components/headers/AcceptRanges'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
        Content-Length:
          $ref: '#/components/headers/ContentLength'
        ETag:
          $ref: '#/components/headers/ETag'
      content:
        application/pdf:
          schema:
            type: string
            format: binary
...